# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HiveOS API — Epic-01 (Organization Bootstrap / v0.1)
  version: 0.2.0
  description: >
    قرارداد حداقلی قابل‌کدنویسی (OpenAPI 3.1) برای endpoint های اپیک ۰۱.
    منبع: متن US-001 تا US-008 در `Dropbox/Hive/doc/Development/epic-01/`.
    استک v0.1 (ADR-019): Python-first (thin API) — این قرارداد توافق یکتای
    بک‌اند، فرانت‌اند و QA است. مسیر پایه `/api/v1` مطابق متن استوری‌ها.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8100
    description: Dev server (v0.1 Python thin API)
security:
  - sessionCookie: []
paths:
...
  /api/v1/ingestion-folder/configure:
    post:
      summary: پیکربندی فولدر Ingestion (US-007)
      description: >
        مسیر فولدر (روی سرور on-premise) را می‌گیرد، دسترسی خواندن را اعتبارسنجی می‌کند و
        سرویس پایش مداوم آن را فعال می‌کند. در نبود پیکربندی معتبر AI model، فعال‌سازی مسدود است.
      operationId: configureIngestionFolder
      tags: [Ingestion]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IngestionFolderConfigure' }
      responses:
        '201':
          description: پایش فعال شد
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionFolderStatus' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403':
          description: بدون دسترسی خواندن / مسیر خارج از محدودهٔ مجاز
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: پیکربندی AI model ناقص/نامعتبر است (FR-010)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/ingestion-folder/status:
    get:
      summary: وضعیت فولدر و پردازش (US-007)
      operationId: getIngestionFolderStatus
      tags: [Ingestion]
      responses:
        '200':
          description: وضعیت پایش و شمارنده‌های اسناد
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionFolderStatus' }
        '404':
          description: فولدر پیکربندی نشده است
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/documents:
    get:
      summary: فهرست صفحه‌بندی‌شدهٔ اسناد (US-007 / S1-18)
      description: >-
        فهرست اسناد شناسایی‌شده با وضعیت پردازش — برای نمایش زنده به Owner.
        پاسخ در پاکت `{items, meta}` برمی‌گردد؛ `meta.total` اندازهٔ کامل
        مجموعهٔ نتیجه پیش از صفحه‌بندی است. sort فقط فیلدهای مجاز را می‌پذیرد.
      operationId: listDocuments
      tags: [Ingestion]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
          description: شمارهٔ صفحه (یک‌مبنا)
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
          description: تعداد آیتم در هر صفحه
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: '-createdAt'
            enum: [createdAt, -createdAt, filename, -filename, sizeBytes, -sizeBytes, status, -status, format, -format]
          description: کلید مرتب‌سازی؛ پیشوند `-` یعنی نزولی. مقدار نامعتبر به `-createdAt` می‌افتد.
        - name: filter
          in: query
          required: false
          schema: { type: string }
          description: جستجوی زیررشته‌ای (غیرحساس به بزرگی) در نام فایل
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [detected, processing, ready, failed]
          description: فیلتر بر اساس وضعیت سند
      responses:
        '200':
          description: فهرست صفحه‌بندی‌شدهٔ اسناد
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentPage' }
  /api/v1/documents/{id}/status:
    get:
      summary: وضعیت یک سند (US-007)
      operationId: getDocumentStatus
      tags: [Ingestion]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: شناسهٔ زمان‌حقیقی (پیشنهادی UUID) سند
      responses:
        '200':
          description: وضعیت
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentStatusResponse' }
        '404':
          description: سند یافت نشد
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: >
        سشن Secure/HttpOnly که پس از US-002 صادر و در US-003 تأیید می‌شود؛
        محدودهٔ Organization را برای درخواست‌های Owner مشخص می‌کند.
  responses:
    BadRequest:
      description: ورودی نامعتبر
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    IngestionFolderConfigure:
      type: object
      additionalProperties: false
      required: [folderPath]
      properties:
        folderPath:
          type: string
          description: مسیر فولدر روی سرور؛ باید درمحدودهٔ مجاز و قابل‌خواندن باشد.
    IngestionFolderStatus:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
        folderPath: { type: [string, 'null'] }
        watchStartedAt: { type: [string, 'null'], format: date-time }
        counts:
          type: object
          properties:
            detected: { type: integer }
            processing: { type: integer }
            ready: { type: integer }
            failed: { type: integer }
    Document:
      type: object
      required: [id, filename, format, status]
      properties:
        id: { type: string }
        filename: { type: string }
        format:
          type: string
          enum: [pdf, docx, txt, md]
        sizeBytes: { type: integer }
        status:
          type: string
          enum: [detected, processing, ready, failed]
        error:
          type: [string, 'null']
          description: در صورت failed دلیل شفاف.
        createdAt: { type: string, format: date-time }
    DocumentStatusResponse:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status:
          type: string
          enum: [detected, processing, ready, failed]
        error: { type: [string, 'null'] }
    PageMeta:
      type: object
      required: [total, page, pageSize, totalPages]
      properties:
        total:
          type: integer
          description: اندازهٔ کامل مجموعهٔ نتیجه پیش از صفحه‌بندی
        page: { type: integer }
        pageSize: { type: integer }
        totalPages:
          type: integer
          description: 'ceil(total / pageSize)'
    DocumentPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Document' }
        meta: { $ref: '#/components/schemas/PageMeta' }
      description: پاکت صفحه‌بندی‌شدهٔ فهرست اسناد (S1-18)
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: کد/پیام خطای کوتاه.
        message:
          type: string
          description: پیام قابل‌فهم برای کاربر.
        auditLogged:
          type: boolean
...